Add files via upload
This is a star sign generator. It takes the users birth month and day and uses these inputs to generate the users star sign from a dictioanry cntaing all the star signs and their meanings.
The final output will be the users star sign and its meaning.
It uses the stars.jpg as the background image.